fix type bug

+++ test_audio.py
@@ -29,7 +29,7 @@ if __name__ == '__main__':
     enhanced_speech = []
     for noisy_slice in noisy_slices:
         z = nn.init.normal(torch.Tensor(1, 1024, 8))
-        noisy_slice = torch.from_numpy(pre_emphasis(noisy_slice[np.newaxis, np.newaxis, :]))
+        noisy_slice = torch.from_numpy(pre_emphasis(noisy_slice[np.newaxis, np.newaxis, :])).type(torch.FloatTensor)
         if torch.cuda.is_available():
             noisy_slice = noisy_slice.cuda()
         noisy_slice = Variable(noisy_slice)
